# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Proyecto Final | Plataforma de Reportes Ciudadanos.
  description: |
    La plataforma de seguridad ciudadana y bienestar comunitario permite a los ciudadanos reportar incidentes en su entorno, facilitando la comunicación y colaboración en tiempo real entre usuarios y moderadores.

    Las APIs están organizadas en los siguientes módulos:
    - Usuarios: Registro, autenticación, activación de cuenta y gestión de perfil.
    - Reportes: Creación, actualización, eliminación y categorización de incidentes (seguridad, emergencias médicas, infraestructura, etc.).
    - Comentarios: Permite a los usuarios aportar información adicional en reportes publicados.
    - Notificaciones: Alertas en tiempo real sobre incidentes cercanos mediante WebSockets, Firebase Cloud Messaging o correo electrónico.
    - Moderadores: Gestión y verificación de reportes, administración de categorías, generación de informes y resolución de incidentes.
    
    Estas APIs garantizan que los usuarios puedan crear, visualizar y gestionar reportes, mientras que los moderadores aseguran la calidad y relevancia de la información compartida en la plataforma.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo (Demo)
...
paths:
...
  /api/usuarios:
    post:
      summary: Registrar un nuevo usuario.
      tags:
        - Usuario
      description: |
        Registra un nuevo usuario en el sistema.
        - Se valida que el correo y el ID no estén repetidos, y que todos los campos sean obligatorios.
        - Se obtiene la ubicación a partir de un mapa interactivo. (Se recomienda usar Mapbox para obtener la ubicación exacta.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Usuario"
      responses:
        "201":
          description: Usuario registrado exitosamente. Código de activación enviado.
        "400":
          description: Error en los datos enviados - Datos faltantes o incorrectos,
        "409":
          description: Conflicto - Email o ID ya existen.
  /api/usuarios/{id}:
    put:
      summary: Actualizar usuario
      tags:
        - Usuario
        - Moderador
      description: |
        Modifica los datos de un usuario existente. 
        - Solo el propietario de la cuenta o un moderador pueden realizar esta acción.
        - Se pueden actualizar la dirección y la ubicación geográfica (latitud y longitud) usando Mapbox.
      parameters:
        - name: id
          in: path
          required: true
          description: Id del usuario a actualizar
          schema:
            type: string
      security:
        - cookieAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: "Nuevo Nombre"
                ciudad:
                  type: string
                  example: "Nueva Ciudad"
                direccion:
                  type: string
                  example: "Nueva direccion"
                ubicacion:
                  $ref: '#/components/schemas/Ubicacion'
      responses:
        "200":
          description: Usuario actualizado exitosamente
        "400":
          description: Error en los datos enviados. Algún campo es inválido o está vacío.
        "401":
          description: No autenticado.
        "403":
          description: Acceso prohibido (ID no coincide con el autenticado).
        "404":
          description: Usuario no encontrado.
        "500":
          description: Error del servidor.

    delete:
      summary: Eliminar usuario
      tags:
        - Usuario
        - Moderador
      description: |
        Borra un usuario del sistema. Solo los administradores o el propio usuario pueden eliminar la cuenta.
      parameters:
        - name: id
          in: path
          required: true
          description: Id del usuario a eliminar
          schema:
            type: string
      security:
        - cookieAuth: [ ]
      responses:
        "204":
          description: Usuario eliminado exitosamente
        "202":
          description: La solicitud fue recibida, pero la eliminación se procesará más tarde (caso de procesamiento asíncrono).
        "400":
          description: El id es inválido o la solicitud está malformada.
        "401":
          description: No se envió token o cookie de autenticación, o es inválido.
        "403":
          description: Acceso prohibido (ID no coincide con el autenticado)
        "404":
          description: Usuario no encontrado
        "500":
          description: Fallo en la base de datos o en el servidor.
        "502":
          description: El servidor no puede comunicarse con otro servicio necesario.

    get:
      summary: Obtener usuario por ID
      tags:
        - Usuario
        - Moderador
      description: |
        Devuelve los datos de un usuario específico mediante su ID. 
        Solo el propio usuario o un moderador pueden acceder a esta información.
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario a obtener
          schema:
            type: string
      security:
        - cookieAuth: [ ]
      responses:
        "200":
          description: Datos del usuario obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "400":
          description: Solicitud incorrecta (ID inválido).
        "401":
          description: No autorizado (cookie faltante o inválida).
        "403":
          description: Acceso prohibido (el usuario autenticado no coincide con el ID).
        "404":
          description: Usuario no encontrado.
        "500":
          description: Error interno del servidor.
  /api/usuarios/{email}:
    get:
      summary: Obtener usuario por Email
      tags:
        - Usuario
        - Moderador
      description: |
        Devuelve los datos de un usuario específico mediante su Email. 
        Solo el propio usuario o un moderador pueden acceder a esta información.
      parameters:
        - name: email
          in: path
          required: true
          description: Email del usuario a obtener
          schema:
            type: string
            format: email
            pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$' # Validación de formato de email
      security:
        - cookieAuth: [ ]
      responses:
        "200":
          description: Datos del usuario obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "400":
          description: Solicitud incorrecta (Email invalido).
        "401":
          description: No autenticado (cookie faltante o inválida).
        "403":
          description: Acceso prohibido (el usuario autenticado no coincide con el email).
        "404":
          description: Usuario no encontrado.
        "500":
          description: Error interno del servidor.
...
  /api/moderadores/categorias/{id}:
    put:
      summary: Editar una categoría de reportes.
      tags:
        - Moderador
      description: |
        Permite a un moderador actualizar el nombre y descripción de una categoría existente.
        - No se permite cambiar una categoría si ya está asociada a reportes.
      security:
        - cookieAuth: [ ]  # Solo moderadores autenticados pueden acceder
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la categoría a modificar.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  description: Nuevo nombre de la categoría.
                descripcion:
                  type: string
                  description: Nueva descripción de la categoría.
      responses:
        "200":
          description: Categoría actualizada exitosamente.
        "400":
          description: Datos inválidos o faltantes.
        "403":
          description: Acceso denegado, solo moderadores pueden editar categorías.
        "404":
          description: Categoría no encontrada.
        "409":
          description: No se puede modificar una categoría asociada a reportes.
        "500":
          description: Error interno del servidor.

    delete:
      summary: Eliminar una categoría de reportes.
      tags:
        - Moderador
      description: |
        Permite a un moderador eliminar una categoría de reportes.
        - No se permite eliminar una categoría si está asociada a algún reporte.
      security:
        - cookieAuth: [ ]  # Solo moderadores autenticados pueden acceder
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la categoría a eliminar.
          schema:
            type: string
      responses:
        "204":
          description: Categoría eliminada exitosamente.
        "403":
          description: Acceso denegado, solo moderadores pueden eliminar categorías.
        "404":
          description: Categoría no encontrada.
        "409":
          description: No se puede eliminar una categoría asociada a reportes.
        "500":
          description: Error interno del servidor.

  # Apis de los reportes
...
  /api/reportes/usuarios/{id}:
    get:
      summary: Obtener todos los reportes de un usuario
      tags:
        - Reporte
        - Usuario
      description: |
        Permite a un usuario autenticado obtener todos los reportes que ha registrado. 
        Un moderador también puede ver los reportes de cualquier usuario.
      security:
        - cookieAuth: [ ]  # Usa autenticación basada en cookies
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario cuyos reportes se desean obtener.
          schema:
            type: string
        - name: estado
          in: query
          required: false
          description: Filtrar reportes por estado (Pendiente, Verificado, Rechazado).
          schema:
            type: string
            enum: [ "Pendiente", "Verificado", "Rechazado" ]
        - name: categoria
          in: query
          required: false
          description: Filtrar reportes por categoría.
          schema:
            type: string
      responses:
        "200":
          description: Lista de reportes obtenida exitosamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reporte"
        "400":
          description: ID de usuario inválido o parámetros incorrectos.
        "401":
          description: No autenticado, se requiere iniciar sesión.
        "403":
          description: Usuario no autorizado para ver estos reportes.
        "404":
          description: No se encontraron reportes para el usuario proporcionado.
        "500":
          description: Error interno del servidor.
...
  /api/favoritos/usuarios/{id}:
    get:
      summary: Obtener la lista de reportes favoritos de un usuario.
      tags:
        - Usuario
      description: |
        Devuelve la lista de reportes que un usuario ha marcado como favoritos.
        - Solo el usuario autenticado puede consultar su propia lista.
        - Se pueden agregar filtros opcionales para mejorar la consulta.
      security:
        - cookieAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario cuya lista de favoritos se desea obtener.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lista de reportes favoritos obtenida exitosamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reporte"
        "400":
          description: Solicitud inválida (ID no válido o error en los parámetros de consulta).
        "401":
          description: Usuario no autenticado.
        "403":
          description: Acceso denegado (un usuario no puede consultar favoritos de otro usuario).
        "404":
          description: Usuario no encontrado.
        "500":
          description: Error interno del servidor.

  #Apis de los comentarios
...
  /api/notificaciones/usuarios/{id}:
    get:
      summary: Obtener todas las notificaciones de un usuario.
      tags:
        - Notificaciones
        - Usuario
      description: |
        Recupera todas las notificaciones asociadas a un usuario específico, incluyendo notificaciones no leídas y leídas. 
           - Solo el usuario autenticado o un moderador pueden consultar estas notificaciones.
           - Las notificaciones incluyen detalles como fecha, estado (leída/no leída).
      security:
        - cookieAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: ID del Usuario cuyas notificaciones se desea consultar.
          schema:
            type: string
        - name: estado
          in: query
          required: false
          description: Filtra por estado de la notificación (leída/no leída).
          schema:
            type: string
            enum: [ "leida", "no_leida" ]
      responses:
        "200":
          description: Lista de notificaciones obtenida exitosamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notificacion"
        "400":
          description: Solicitud inválida (por ejemplo, ID no válido o parámetro incorrecto).
        "401":
          description: Usuario no autenticado.
        "403":
          description: Acceso denegado (un usuario no puede consultar notificaciones de otro usuario).
        "404":
          description: Usuario no encontrado.
        "500":
          description: Error en el servidor.
...
components:
  schemas:
...
    Usuario:
      type: object
      properties:
        nombre:
          type: string
          example: "Nicolas Cabrera"
        direccion:
          type: string
          example: "Calle 5 Norte #12-34"
        ubicacion:
          $ref: '#/components/schemas/Ubicacion'
        ciudad:
          type: string
          example: "Bogota"
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            estado:
              type: string
              enum: [ "Activo","Inactivo","Eliminado" ]
              example: "Inactivo"
            rol:
              type: string
              enum: [ "Moderador","Usuario","Megatron" ]
              example: "Usuario"
            password:
              type: string
              example: "asdac"
            email:
              type: string
              format: email
              pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$' # Formato obligatorio
              example: "nicolasC@example.com"
            reportes:
              type: array
              nullable: true
              items:
                $ref: '#/components/schemas/Reporte'
            reportesFav:
              type: array
              nullable: true
              items:
                $ref: '#/components/schemas/Reporte'
            notificaciones:
              type: array
              nullable: true
              items:
                $ref: '#/components/schemas/Notificacion'


    #Esquema del reporte.
    Reporte:
      type: object
      properties:
        Id:
          type: string
          format: uuid
          example: "AB20"
        IdUsuario:
          type: string
          format: uuid
          example: "1032"
        titulo:
          type: string
          example: "Reporte de Nicolas"
        categoria:
          $ref: '#/components/schemas/Categoria'
        estaResuelto:
          type: string
          enum: ["Resuelto","NoResuelto"]
          example: "NoResuelto"
        descripcion:
          type: string
          format: descripción
          example: "Problemas de robos y violencia"
        estadoReporte:
          type: string
          enum: ["Pendiente", "Verificado", "Rechazado"]
          example: "Verificado"
        ubicacion:
          $ref: '#/components/schemas/Ubicacion'
        foto:
          type: string
          format: uri
          example: "https://example.com/foto-reporte.jpg"
        numImportancia:
          type: integer
          description: Número de usuarios que han calificado este reporte como importante
          example: 0
        severidad:
          type: string
          description: Nivel de severidad basado en la cantidad de votos
          enum: [ "Baja", "Media", "Alta" ]
          example: "Media"
        comentarios:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Comentario'
        historialEstado:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Historial'

    #Esquema de los comentarios.
    Comentario:
      type: object
      properties:
        idComentario:
          type: string
          format: uuid
          example: 101
        idUsuario:
          type: string
          format: uuid
          example: "1032"
        idReporte:
          type: string
          format: uuid
        contenido:
          type: string
          example: "Tu comentario es pura falsedad"
        fecha:
          type: string
          format: data-time
          example: "2025-02-18T14:30:00Z"
        comentarios:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Comentario'

    # Esquema del Historial.
    Historial:
      type: object
      properties:
        estadoAnterior:
          type: string
          description: "Estado previo del reporte."
        nuevoEstado:
          type: string
          description: "Estado actualizado del reporte."
        fechaCambio:
          type: string
          format: date-time
          description: "Fecha y hora del cambio de estado."
        moderadorId:
          type: string
          description: "ID del moderador que realizó el cambio."
        motivo:
          type: string
          description: "Motivo del cambio de estado (opcional)."

    #Esquema de categoría.
    Categoria:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 2A1
        nombre:
          type: string
          example: "Infraestructura"
        descripcion:
          type: string
          example: "Calles en mal estado"

    # Esquema de Ubicación
    Ubicacion:
      type: object
      properties:
        latitud:
          type: number
          format: float
          example: -34.6037
        longitud:
          type: number
          format: float
          example: -58.3816

    # Esquema de notificación
    Notificacion:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: "ID único de la notificación"
        idDestinatario:
          type: string
          format: uuid
          description: "ID del usuario que recibe la notificación"
        idReporte:
          type: string
          format: uuid
          description: ID del reporte asociado a la notificación.
        titulo:
          type: string
          description: Título de la notificación.
        mensaje:
          type: string
          description: "Mensaje de la notificación"
        fecha:
          type: string
          format: date-time
          description: "Fecha y hora en que se envió la notificación"
        leida:
          type: string
          enum: ["Leido","Sin_Leer"]
          example: "Leido"
          description: "Indica si la notificación ha sido vista por el usuario"
        ubicacionReporte:
          description: "Coordenadas del reporte asociado."
          items:
            $ref: '#/components/schemas/Ubicacion'
        radio:
          type: number
          format: float
          description: Radio en kilómetros para notificar usuarios cercanos.

    # Esquema para código verificación
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSIONID  # Cookie de autenticación con la sesión del usuario



# Firma ROT5:  QwE Pjanfyfr(Nicolas Cabrera Serrano)